+++ Modules/31-BitLocker.ps1
@@ -61,6 +61,7 @@ function Show-BitLockerManagement {
     }
 
     while ($true) {
+        if ($global:ReturnToMainMenu) { return }
         # Show current BitLocker status
         Clear-Host
         Write-OutputColor "" -color "Info"